backend/requirements.txt
fastapi
pydantic
uvicorn
rich
python-dotenv
bcrypt
requests
aiohttp
aiodns
motor
tornado
click
pytz
pandas
openpyxl
APScheduler
python-multipart
firebase-admin
pyarrow
psutil
pydub
ffmpeg-python
gitpython